Move files to common folder
