+++ WebApiSeed/Common/IServiceResult.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WebApiSeed.Dtos
+namespace WebApiSeed.Common
 {
     public interface IServiceResult
     {
