+++ WebApiSeed/Common/ServiceResult.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using WebApiSeed.Dtos;
 
-namespace WebApiSeed.Dtos
+namespace WebApiSeed.Common
 {
     public class ServiceResult<TDto, TServiceResult>
         where TDto : BaseDto
